Project-2-API-Integration/requirements.txt
requests==2.28.1
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
pytest==8.3.4
pyyaml==6.0
